Move back DepositShare to it's own file
